# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TreasureTots Creations e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: subcategory
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /v1/products/search:
    get:
      operationId: searchProducts
      tags: [products]
      summary: Search products
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /v1/products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /v1/cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /v1/orders/initialize:
    post:
      operationId: initializeOrder
      tags: [orders]
      summary: Initialize order and create Razorpay order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInitInput"
      responses:
        "201":
          description: Order initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderInitResponse"
  /v1/orders/verify:
    post:
      operationId: verifyPayment
      tags: [orders]
      summary: Verify Razorpay payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /v1/orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        stock:
          type: integer
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        slug:
          type: string
        isBuyable:
          type: boolean
        isActive:
          type: boolean
      required: [id, name, coverImage, images, category, slug, isBuyable, isActive]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer
      required: [products, total, page, perPage]
...
    CartItem:
      type: object
      properties:
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        childName:
          type: ["string", "null"]
      required: [productId, product, quantity]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required: [items, total, itemCount]
    CartItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        childName:
          type: ["string", "null"]
      required: [productId]
    CartItemUpdate:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]
    Address:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        houseNo:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required: [id, fullName, phone, houseNo, street, city, state, pincode]
    AddressInput:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        houseNo:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required: [fullName, phone, houseNo, street, city, state, pincode]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
      required: [id, productId, quantity, price]
    Order:
      type: object
      properties:
        id:
          type: integer
        totalAmount:
          type: number
        paymentStatus:
          type: string
        orderStatus:
          type: string
        childName:
          type: ["string", "null"]
        shippingAddress:
          $ref: "#/components/schemas/Address"
        razorpayOrderId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required: [id, totalAmount, paymentStatus, orderStatus, items, createdAt]
    OrderInitInput:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required: [productId, quantity]
        childName:
          type: ["string", "null"]
        address:
          $ref: "#/components/schemas/AddressInput"
      required: [items, address]
    OrderInitResponse:
      type: object
      properties:
        orderId:
          type: integer
        razorpayOrderId:
          type: string
        amount:
          type: number
        key:
          type: string
      required: [orderId, razorpayOrderId, amount, key]
    PaymentVerifyInput:
      type: object
      properties:
        razorpayPaymentId:
          type: string
        razorpayOrderId:
          type: string
        razorpaySignature:
          type: string
        orderId:
          type: integer
      required: [razorpayPaymentId, razorpayOrderId, razorpaySignature, orderId]
...