# lib/api-zod/src/generated/types/addressInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AddressInput {
  fullName: string;
  phone: string;
  houseNo: string;
  street: string;
  city: string;
  state: string;
  pincode: string;
  isDefault?: boolean;
}