# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartItem {
  productId: number;
  product: Product;
  quantity: number;
  /** @nullable */
  childName?: string | null;
}