# lib/api-zod/src/generated/types/orderInitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { AddressInput } from './addressInput';
import type { OrderInitInputItemsItem } from './orderInitInputItemsItem';

export interface OrderInitInput {
  items: OrderInitInputItemsItem[];
  /** @nullable */
  childName?: string | null;
  address: AddressInput;
}