# lib/api-zod/src/generated/types/orderInitResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInitResponse {
  orderId: number;
  razorpayOrderId: string;
  amount: number;
  key: string;
}